requirements.txt
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.3